pyproject.toml
[project]
name = "farhaan-bukh-sh"
version = "0.1.0"
description = "Static site generator for farhaan.bukh.sh"
requires-python = ">=3.13"
dependencies = [
    "jinja2>=3.1",
    "pyyaml>=6.0",
    "markdown>=3.5",
]
